more centralization of common elements
