+++ webng/analysis/analysis.py
@@ -1,7 +1,19 @@
-import sys
+import sys, os
 
 class weAnalysis:
-    def __init__(self):
+    def __init__(self, opts):
+        # get this in so we can import system.py if need be
+        sys.path.append(opts["sim_name"])
+        # keep opts around
+        self.opts = opts
+        # Set work path
+        self.work_path = self._getd(opts, "work-path", default=os.getcwd(), required=False)
+        # we want to go there
+        assert os.path.isdir(self.work_path), "Work path: {} doesn't exist".format(self.work_path)
+        self.curr_path = os.getcwd()
+        os.chdir(self.work_path)
+        # Set names if we have them
+        self.set_names(opts["pcoords"])
         # we need to write assignment.py sometimes
         self.pull_data_str =  "import numpy as np\n"
         self.pull_data_str += "def pull_data(n_iter, iter_group):\n"
@@ -21,3 +33,12 @@ class weAnalysis:
         if required and (val is None):
             sys.exit("{} is not specified in the dictionary".format(key))
         return val
+    
+    def set_names(self, names):
+        if names is not None:
+            self.names = dict( zip(range(len(names)), names) )
+        else:
+            # We know the dimensionality, can assume a 
+            # naming scheme if we don't have one
+            print("Giving default names to each dimension")
+            self.names = dict( (i, str(i)) for i in range(self.dims) )
